refactor: ReservationApiController 도메인별 분리
