+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,48 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationResponse;
+import roomescape.dto.SaveReservationRequest;
+import roomescape.service.ReservationService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class ReservationController {
+
+    private final ReservationService reservationService;
+
+    public ReservationController(final ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping("/reservations")
+    public List<ReservationResponse> getReservations() {
+        return reservationService.getReservations()
+                .stream()
+                .map(ReservationResponse::from)
+                .toList();
+    }
+
+    @PostMapping("/reservations")
+    public ResponseEntity<ReservationResponse> saveReservation(@RequestBody final SaveReservationRequest request) {
+        final Reservation savedReservation = reservationService.saveReservation(request);
+
+        return ResponseEntity.created(URI.create("/reservations/" + savedReservation.getId()))
+                .body(ReservationResponse.from(savedReservation));
+    }
+
+    @DeleteMapping("/reservations/{reservation-id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable("reservation-id") final Long reservationId) {
+        reservationService.deleteReservation(reservationId);
+        return ResponseEntity.noContent().build();
+    }
+}
