+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,48 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.SaveReservationTimeRequest;
+import roomescape.service.ReservationService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class ReservationTimeController {
+
+    private final ReservationService reservationService;
+
+    public ReservationTimeController(final ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping("/times")
+    public List<ReservationTimeResponse> getReservationTimes() {
+        return reservationService.getReservationTimes()
+                .stream()
+                .map(ReservationTimeResponse::from)
+                .toList();
+    }
+
+    @PostMapping("/times")
+    public ResponseEntity<ReservationTimeResponse> saveReservationTime(@RequestBody final SaveReservationTimeRequest request) {
+        final ReservationTime savedReservationTime = reservationService.saveReservationTime(request);
+
+        return ResponseEntity.created(URI.create("/times/" + savedReservationTime.getId()))
+                .body(ReservationTimeResponse.from(savedReservationTime));
+    }
+
+    @DeleteMapping("/times/{reservation-time-id}")
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable("reservation-time-id") final Long reservationTimeId) {
+        reservationService.deleteReservationTime(reservationTimeId);
+        return ResponseEntity.noContent().build();
+    }
+}
